+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Update aria-label for action menu buttons based on context. Refs UIIN-3685.
 * Revert UIIN-3377. Make submitting an Item Form default to Save & close behaviour. Refs UIIN-3596.
 * Add instance custom links settings. Refs UIIN-3652.
+* Add custom links to instance action menu. Refs UIIN-3653.
 
 ## [14.0.6](https://github.com/folio-org/ui-inventory/tree/v14.0.6) (2026-08-12)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v14.0.5...v14.0.6)